feat: adjust edit workflow (#136)
* feat: adjust edit workflow

* chore: remove unused varibles
